Embeddings initialization refactor
